Add test for auto encrypted publish
